+++ changelog.md
@@ -309,3 +309,5 @@
 * Fixed zombies not always using zombie clothing
 * Villager on fire will now burn their clothes
 * The Sirben cult appeared
+* Added 50.000 names from 55 different countries
+  * Config option available to use modern USA names only